requirements.txt
asgiref==3.4.1
CacheControl==0.12.6
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
djangorestframework==3.12.4
idna==3.3
jsonpickle==2.0.0
msgpack==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
weatherapi==0.0.4
